add typescript support to e2e tests
