+++ e2e/cypress/package.json
@@ -1,8 +1,22 @@
 {
   "devDependencies": {
     "@cypress/request": "2.88.7",
+    "@mattermost/types": "^7.0.0",
     "@testing-library/cypress": "7.0.6",
+    "@types/async": "^3.2.13",
+    "@types/authenticator": "^1.1.1",
+    "@types/express": "^4.17.13",
+    "@types/fs-extra": "^9.0.13",
     "@types/lodash": "4.14.175",
+    "@types/lodash.intersection": "^4.4.7",
+    "@types/lodash.mapkeys": "^4.6.7",
+    "@types/lodash.without": "^4.4.7",
+    "@types/mime-types": "^2.1.1",
+    "@types/mochawesome": "^6.2.1",
+    "@types/pdf-parse": "^1.1.1",
+    "@types/recursive-readdir": "^2.2.1",
+    "@types/shelljs": "^0.8.11",
+    "@types/uuid": "^8.3.4",
     "async": "3.2.1",
     "authenticator": "1.1.5",
     "aws-sdk": "2.1002.0",
@@ -45,19 +59,21 @@
     "recursive-readdir": "2.2.2",
     "shelljs": "0.8.4",
     "timezones.json": "1.5.3",
+    "typescript": "^4.7.4",
     "uuid": "8.3.2",
     "yargs": "17.2.1"
   },
   "scripts": {
+    "check-types": "tsc -b",
     "cypress:open": "cross-env TZ=Etc/UTC cypress open",
-    "cypress:run": "cross-env TZ=Etc/UTC cypress run --browser chrome --config ignoreTestFiles='**/enterprise/**/*.js'",
-    "cypress:run:enterprise": "cross-env TZ=Etc/UTC cypress run --browser chrome --spec='tests/integration/enterprise/**/*_spec.js'",
-    "cypress:run:firefox": "cross-env TZ=Etc/UTC cypress run --browser firefox --config ignoreTestFiles='**/enterprise/**/*.js'",
-    "cypress:run:firefox:enterprise": "cross-env TZ=Etc/UTC cypress run --browser firefox --spec='tests/integration/enterprise/**/*_spec.js'",
-    "cypress:run:edge": "cross-env TZ=Etc/UTC cypress run --browser edge --config ignoreTestFiles='**/enterprise/**/*.js'",
-    "cypress:run:edge:enterprise": "cross-env TZ=Etc/UTC cypress run --browser edge --spec='tests/integration/enterprise/**/*_spec.js'",
-    "cypress:run:electron": "cross-env TZ=Etc/UTC cypress run --browser electron --config ignoreTestFiles='**/enterprise/**/*.js'",
-    "cypress:run:electron:enterprise": "cross-env TZ=Etc/UTC cypress run --browser electron --spec='tests/integration/enterprise/**/*_spec.js'",
+    "cypress:run": "cross-env TZ=Etc/UTC cypress run --browser chrome --config ignoreTestFiles='**/enterprise/**/*.{js,ts}'",
+    "cypress:run:enterprise": "cross-env TZ=Etc/UTC cypress run --browser chrome --spec='tests/integration/enterprise/**/*_spec.{js,ts}'",
+    "cypress:run:firefox": "cross-env TZ=Etc/UTC cypress run --browser firefox --config ignoreTestFiles='**/enterprise/**/*.{js,ts}'",
+    "cypress:run:firefox:enterprise": "cross-env TZ=Etc/UTC cypress run --browser firefox --spec='tests/integration/enterprise/**/*_spec.{js,ts}'",
+    "cypress:run:edge": "cross-env TZ=Etc/UTC cypress run --browser edge --config ignoreTestFiles='**/enterprise/**/*.{js,ts}'",
+    "cypress:run:edge:enterprise": "cross-env TZ=Etc/UTC cypress run --browser edge --spec='tests/integration/enterprise/**/*_spec.{js,ts}'",
+    "cypress:run:electron": "cross-env TZ=Etc/UTC cypress run --browser electron --config ignoreTestFiles='**/enterprise/**/*.{js,ts}'",
+    "cypress:run:electron:enterprise": "cross-env TZ=Etc/UTC cypress run --browser electron --spec='tests/integration/enterprise/**/*_spec.{js,ts}'",
     "start:webhook": "node webhook_serve.js",
     "test": "start-test start:webhook 3000 test:ci",
     "test:ci": "node run_tests.js",
