--- e2e/cypress/tests/support/api/setup.js
@@ -1,48 +1,0 @@
-// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
-// See LICENSE.txt for license information.
-
-Cypress.Commands.add('apiInitSetup', ({
-    loginAfter = false,
-    promoteNewUserAsAdmin = false,
-    hideAdminTrialModal = true,
-    userPrefix,
-    teamPrefix = {name: 'team', displayName: 'Team'},
-    channelPrefix = {name: 'channel', displayName: 'Channel'},
-} = {}) => {
-    return cy.apiCreateTeam(teamPrefix.name, teamPrefix.displayName).then(({team}) => {
-        // # Add public channel
-        return cy.apiCreateChannel(team.id, channelPrefix.name, channelPrefix.displayName).then(({channel}) => {
-            return cy.apiCreateUser({prefix: userPrefix || (promoteNewUserAsAdmin ? 'admin' : 'user')}).then(({user}) => {
-                if (promoteNewUserAsAdmin) {
-                    cy.apiPatchUserRoles(user.id, ['system_admin', 'system_user']);
-
-                    // Only hide start trial modal for admin since it's not applicable to other users
-                    cy.apiSaveStartTrialModal(user.id, hideAdminTrialModal.toString());
-                }
-
-                return cy.apiAddUserToTeam(team.id, user.id).then(() => {
-                    return cy.apiAddUserToChannel(channel.id, user.id).then(() => {
-                        const getUrl = (channelName) => `/${team.name}/channels/${channelName}`;
-
-                        const data = {
-                            channel,
-                            team,
-                            user,
-                            channelUrl: getUrl(channel.name),
-                            offTopicUrl: getUrl('off-topic'),
-                            townSquareUrl: getUrl('town-square'),
-                        };
-
-                        if (loginAfter) {
-                            return cy.apiLogin(user).then(() => {
-                                return cy.wrap(data);
-                            });
-                        }
-
-                        return cy.wrap(data);
-                    });
-                });
-            });
-        });
-    });
-});
