--- e2e/cypress/tests/tsconfig.json
@@ -1,10 +1,0 @@
-{
-    "compilerOptions": {
-        "allowJs": true,
-        "baseUrl": "../node_modules",
-        "types": [
-            "cypress",
-        ]
-    },
-    "include": ["**/*.*"]
-}
